+++ Java/Concentus/src/main/java/org/concentus/DecodeAPI.java
@@ -149,7 +149,7 @@ class DecodeAPI {
         psDec.nChannelsAPI = decControl.nChannelsAPI;
         psDec.nChannelsInternal = decControl.nChannelsInternal;
 
-        if (decControl.API_sampleRate > (int) SilkConstants.MAX_API_FS_KHZ * 1000 || decControl.API_sampleRate < 8000) {
+        if (decControl.API_sampleRate > SilkConstants.MAX_API_FS_KHZ * 1000 || decControl.API_sampleRate < 8000) {
             ret = SilkError.SILK_DEC_INVALID_SAMPLING_FREQUENCY;
             return (ret);
         }
